requirements.txt
requests==2.13.0
PyYAML==3.12
boto3==1.4.4
botocore==1.5.36
s3transfer==0.1.10
